requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
pydantic-settings
pypdf
python-docx
